Refactor: Food 모듈 분리 (#26)  - Food모듈 분리
