+++ app/database_model.py
@@ -2,3 +2,4 @@ import modules.user.infra.db_models.user
 import modules.track.infra.db_models.track
 import modules.mealday.infra.db_models.mealday
 import modules.track.infra.db_models.track_participant
+import modules.food.infra.db_models.food
