+++ app/modules/food/interface/schema/food_schema.py
@@ -70,6 +70,23 @@ class UpdateDishBody(BaseModel):
     track_goal: Optional[bool] = None
     size: Optional[float] = None
 
+class Food_Data(BaseModel):
+    label: Optional[int] = 0
+    name: Optional[str] = "음식명"
+    size: Optional[float] = 0.0
+    calorie: Optional[float] = 0.0
+    carb: Optional[float] = 0.0
+    sugar: Optional[float] = 0.0
+    fat: Optional[float] = 0.0
+    protein: Optional[float] = 0.0
+    calcium: Optional[float] = 0.0
+    phosphorus: Optional[float] = 0.0
+    sodium: Optional[float] = 0.0
+    magnesium: Optional[float] = 0.0
+    iron: Optional[float] = 0.0
+    zinc: Optional[float] = 0.0
+    cholesterol: Optional[float] = 0.0
+    trans_fat: Optional[float] = 0.0
 
 class Dish_Full(BaseModel):
     id: str
